코인 상승문제 해결

+++ src/KMUbit/Coin.java
@@ -22,7 +22,7 @@ public class Coin {
 	public static void tomorrow() {
 
 		for (int i = 0; i < 6; i++) { // 다음 날 가격 랜덤 설정
-			int a = (int) Math.random() * 30;
+			double a = Math.random() * 30;
 			double b = 0;
 			if (a <= 10) // 3분의 1 확률로
 			{
@@ -51,8 +51,7 @@ public class Coin {
 
 			}
 
-			if (a > 26 & a <= 30) // 30분의 2 확률로
-			{
+			if (a > 26 & a <= 30) {
 				do {
 					b = Math.random() * 300 + 500;
 					tm_price[i] = (int) (price[i] - b); // 원래 가격- 500 ~ 800
